navbar css refactoring + renamed and moved ad service
