+++ src/app/shared/components/navbar/navbar.component.scss
@@ -21,11 +21,8 @@
 }
 
 .mobile-curved-background {
-  position: absolute;
-  background-color: $tertiary;
-  width: 100%;
+  @extend .mobile-top-nav-background;
   height: 139%;
-  top: 0;
   transform: rotate(-22deg);
   border-bottom-left-radius: 575px 427px;
   border-bottom-right-radius: 372px 18px;
@@ -94,11 +91,9 @@
 }
 
 #account-menu-items-wrapper {
+  @extend .desktop-navbar-icon;
   height: 84vh;
-  display: flex;
   justify-content: space-evenly;
-  align-items: center;
-  flex-direction: column;
   background-color: $tertiary;
 }
 
@@ -176,9 +171,7 @@ ul#desktop-navbar-icons {
 }
 
 #desktop-account-icon {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
+  @extend .desktop-navbar-icon;
   text-decoration: none;
   color: black;
   padding-left: 45px;
